phone_calls: replace is-customer with negation (#61)
## What is the goal of this PR?
Prior to introducing negation in Graql, we had an attribute called `is_customer` assigned to `person` in the `phone_calls` schema. This was done as a workaround in the absence of negation. Now the `phone_calls` example makes use of negation to query for people who are _not_ customers, within the `phone_calls` knowledge graph.

## What are the changes implemented in this PR?
- closes #59 
- removes the definition of `is_customer` attribute and its assignment to `person` from the `phone-calls-schema.gql`
- updates the migration scripts to not assign a value for the `is-customer` attribute when construction the insert queries.
- updates the single one query that used the `is-customer` attribute to use a negation instead.
- resolves issues with the `queries.js` that were and still are not covered by `test.js`. there is an open issue (https://github.com/graknlabs/examples/issues/60) to attend to this.