Introduce an example using TypeDB-Loader with the Catalogue of Life dataset (#119)
## What is the goal of this PR?

We add an example using TypeDB-Loader. The example uses the [Catalogue of Life](https://www.catalogueoflife.org/) dataset to represent the taxonomic structure and some horizontal relations (e.g. geographical distribution). We aim to preserve as much data as possible during migration, within reason.

## What are the changes implemented in this PR?

`Loader` fetches the dataset data off the web (~500 MB compressed), inflates and prepares it for typedb-loader to process and upload into the catalogue-of-life database. The schema includes rules that should make for interesting inference examples.

Order of operations:
* (optionally) fetch and unzip the data;
* split Distributions.tsv based on entry format (`prepareData()`), so it can be loaded in two passes for two different kinds of entity (marine region by mrgid or description);
* populate the typedb-loader config and extract the database schema to the filesystem;
* pass the baton to typedb-loader for it to do its thing—namely, load the data.