GitHub Data Explorer (#127)
## What is the goal of this PR?

We have created an example application that allows demonstrates the ability to dynamically download, migrate and query data using TypeDB, TypeQL with an interface written using Compose Multiplatform. Users can download and run this example, as well as read the source code to see how the end result was achieved.

This PR does not remove any existing functionality.

## What are the changes implemented in this PR?

We introduce a github folder that contains all relevant data including folders for the datasets and schema. Some extra data is provided in the screenshots/, schemas/ and datasets/ folders.

The logic of the application is split across the state/ and view/ folders.
- view/ contains only GitApplication, which is responsible for the GUI and is driven by Compose.
- state/ contains several files:
    * Downloader, which is responsible for downloading repositories from GitHub and writing them to the datasets folder.
    * Explorer, which is responsible for opening a connection to TypeDB and querying the dataset using the given query functions.
    * Migrator, which is responsible for clearing the existing database and importing all of the data created by Downloader.
    * GlobalState, which manages the global state of the program for view/GitApplication.
    * Interfaces, which provides the toJson and fromJson interfaces.
    * Github, which provides classes, representing a subset of the information that GitHub's API gives to us, that implement the toJson and fromJson interfaces.

This structure is typical for a Compose application.